requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch 
torchaudio 
soundfile 
omegaconf
epitran 
audioread
requests
dtwalign
eng_to_ipa
pandas
flask
flask_cors
pickle-mixin
sqlalchemy
transformers
sentencepiece
transformers
g2p_en
langchain==0.2.0
openai-whisper==20231117
pygame==2.5.2
gtts==2.5.1
pyaudio==0.2.14
python-dotenv==1.0.1
langchain-groq==0.1.4
scipy==1.13.1
uvicorn
fastapi
edge-tts
pydantic>=2.5.2
pydub
langdetect
better-profanity
sentence_transformers
nltk